backend/requirements.txt
fastapi
uvicorn
pydantic_settings
sqlalchemy
passlib
python-easy-rsa
easyrsa
cryptography
psycopg2-binary
pydantic[email]
python-jose[cryptography]
python-multipart